Commite cliente controller

+++ src/main/java/service/ServiceCliente.java
@@ -7,8 +7,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.yaml.snakeyaml.tokens.Token.ID;
 
+import exeptions.ClienteExeptions;
 import model.Cliente;
-import model.Compra;
 import repository.ClienteREPO;
 
 @Service
@@ -18,21 +18,29 @@ public class ServiceCliente {
 	private ClienteREPO clienteRepo;
 
 	public Cliente criarCliente(LocalDate data_nascimento, String cpf, String nome, String phone, char sexo,
-			String email) throws Exception {
-		Cliente cliente = new Cliente(data_nascimento, cpf, nome, phone, sexo, email);
-		return cadastroCliente(cliente);
+			String email) throws ClienteExeptions {
+		try {
+			Cliente cliente = new Cliente(data_nascimento, cpf, nome, phone, sexo, email);
+			return cadastroCliente(cliente);
+		} catch (ClienteExeptions e) {
+			throw new ClienteExeptions("Tipo de dado invalido ou fora de ordem");
+		}
 
 	}
 
-	public Cliente cadastroCliente(Cliente cliente) throws Exception {
+	public Cliente cadastroCliente(Cliente cliente) throws ClienteExeptions {
 		if (clienteRepo.existsByCpf(cliente.getCpf())) {
-			throw new Exception("Cliente com CPF já cadastrado.");
+			throw new ClienteExeptions("Cliente com CPF já cadastrado.");
 		}
 		return clienteRepo.save(cliente);
 	}
 
-	public void deleteCliente(Cliente cliente) {
-		clienteRepo.delete(cliente);
+	public void deleteCliente(Cliente cliente) throws ClienteExeptions {
+		if (clienteRepo.existsByCpf(cliente.getCpf())) {
+			clienteRepo.delete(cliente);
+		} else {
+			throw new ClienteExeptions("Cliente Não encontrado");
+		}
 	}
 
 	public Iterable<Cliente> getAllClientes() {
